Personal Story Json and Name File
NPCs now have randomised names (from a txt file) and randomised story arcs (from a json file).
I've added the data files in Assets -> JSON and the script that loads everything is on the main scene (sits on JsonLoader gameobject)

FYI if you put a newline at the end of the name text file the empty line becomes a blank name. I was gonna fix it but it's like 1:30am